Separate script to build blockchain binaries
